requirements.txt
discord.py==2.3.2
httpx==0.27.0
aiosqlite==0.20.0
python-dotenv==1.0.1
